backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.3.0
PyJWT==2.10.1
pymongo==4.14.0
python-dotenv==1.1.1
Werkzeug==3.1.3
